requirements.txt
python-dotenv>=0.20.0
pyyaml>=6.0
ccxt>=4.0.0
pandas>=1.5.0
pandas-ta>=0.3.14b
numpy>=1.23.0
python-telegram-bot>=20.0,<21.0
pytz>=2022.6